directory stucture refactor
